пр11. v1.00

+++ src/utils/Api.js
@@ -18,7 +18,7 @@ class Api {
     })
       .then((res) => this._getResponseData(res))
   }
-  addNewCard(name, link) {
+  addNewCard({name, link}) {
     return fetch(`${this._url}/cards`, {
       method: "POST",
       headers: this._headers,
@@ -58,24 +58,25 @@ class Api {
       method: "PATCH",
       headers: this._headers,
       body: JSON.stringify({
-        avatar: data.avatarUrlInput,
+        avatar: data.avatar,
       })
     })
       .then((res) => this._getResponseData(res))
   }
-  likeCard(cardId) {
-    return fetch(`${this._url}cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    })
-      .then((res) => this._getResponseData(res))
-  }
-  removeLikeCard(cardId) {
-    return fetch(`${this._url}cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    })
-      .then((res) => this._getResponseData(res))
+  changeLikeCardStatus(cardId, isLiked) {
+    if (!isLiked) {
+      return fetch(`${this._url}cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      })
+        .then((res) => this._getResponseData(res))
+    } else {
+      return fetch(`${this._url}cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      })
+        .then((res) => this._getResponseData(res))
+    }
   }
 }
